+++ Encore/src/users/player.h
@@ -12,20 +12,18 @@
 #include "song/scoring.h"
 // #include "libstud-uuid/uuid/uuid.hxx"
 
-
-
 class Band {
-/**
- * @brief Do Not Use. Outdated.
- */
+    /**
+     * @brief Do Not Use. Outdated.
+     */
     std::filesystem::path ScoreFile;
     bool SoloGameplay = true; // to be true until multiple players
 };
 
 class PlayerGameplayStats {
-/**
- * @brief Statistics/statistics manager for individual players during gameplay
- */
+    /**
+     * @brief Statistics/statistics manager for individual players during gameplay
+     */
 public:
     PlayerGameplayStats();
 
@@ -63,11 +61,11 @@ public:
     bool Miss = false;
     int ScoreToDisplay();
 
-    std::vector<float> drumSmasherRotations = {0,0,0,0};
-    std::vector<float> drumSmasherHeights = {0,0,0,0};
+    std::vector<float> drumSmasherRotations = { 0, 0, 0, 0 };
+    std::vector<float> drumSmasherHeights = { 0, 0, 0, 0 };
 
-    std::vector<float> fiveLaneSmasherRotation = {0,0,0,0,0};
-    std::vector<float> fiveLaneSmasherHeights = {0,0,0,0,0};
+    std::vector<float> fiveLaneSmasherRotation = { 0, 0, 0, 0, 0 };
+    std::vector<float> fiveLaneSmasherHeights = { 0, 0, 0, 0, 0 };
 
     std::vector<bool> HeldFrets { false, false, false, false, false };
     std::vector<bool> HeldFretsAlt { false, false, false, false, false };
@@ -147,27 +145,27 @@ public:
     float comboFillCalc();
 };
 
-#define PLAYER_JSON_SETTINGS                                    \
-    SETTING_ACTION(int,     Difficulty,         "diff")         \
-    SETTING_ACTION(int,     Instrument,         "inst")         \
-    SETTING_ACTION(float,   InputCalibration,   "inputOffset")  \
-    SETTING_ACTION(float,   NoteSpeed,          "NoteSpeed")    \
-    SETTING_ACTION(bool,    ProDrums,           "proDrums")     \
-    SETTING_ACTION(bool,    Bot,                "bot")          \
-    SETTING_ACTION(float,   HighwayLength,      "length")       \
-    SETTING_ACTION(bool,    ClassicMode,        "classic")      \
-    SETTING_ACTION(bool,    LeftyFlip,          "lefty")
+#define PLAYER_JSON_SETTINGS                                                             \
+    SETTING_ACTION(int, Difficulty, "diff")                                              \
+    SETTING_ACTION(int, Instrument, "inst")                                              \
+    SETTING_ACTION(float, InputCalibration, "inputOffset")                               \
+    SETTING_ACTION(float, NoteSpeed, "NoteSpeed")                                        \
+    SETTING_ACTION(bool, ProDrums, "proDrums")                                           \
+    SETTING_ACTION(bool, Bot, "bot")                                                     \
+    SETTING_ACTION(float, HighwayLength, "length")                                       \
+    SETTING_ACTION(bool, ClassicMode, "classic")                                         \
+    SETTING_ACTION(bool, LeftyFlip, "lefty")
 
-#define PLAYER_CONFIG_LIST                          \
-    SETTING_ACTION(int,     Difficulty)             \
-    SETTING_ACTION(int,     Instrument)             \
-    SETTING_ACTION(float,   InputCalibration)       \
-    SETTING_ACTION(float,   NoteSpeed)              \
-    SETTING_ACTION(bool,    ProDrums)               \
-    SETTING_ACTION(bool,    Bot)                    \
-    SETTING_ACTION(float,   HighwayLength)          \
-    SETTING_ACTION(bool,    ClassicMode)            \
-    SETTING_ACTION(bool,    LeftyFlip)
+#define PLAYER_CONFIG_LIST                                                               \
+    SETTING_ACTION(int, Difficulty)                                                      \
+    SETTING_ACTION(int, Instrument)                                                      \
+    SETTING_ACTION(float, InputCalibration)                                              \
+    SETTING_ACTION(float, NoteSpeed)                                                     \
+    SETTING_ACTION(bool, ProDrums)                                                       \
+    SETTING_ACTION(bool, Bot)                                                            \
+    SETTING_ACTION(float, HighwayLength)                                                 \
+    SETTING_ACTION(bool, ClassicMode)                                                    \
+    SETTING_ACTION(bool, LeftyFlip)
 
 class Player {
     /**
@@ -190,7 +188,7 @@ public:
     int joypadID;
     bool ReadiedUpBefore;
     bool Online;
-    int ActiveSlot{};
+    int ActiveSlot {};
 
     void ResetGameplayStats();
 
@@ -203,7 +201,6 @@ public:
     // NOTE! this is only for like. local information and
     // not actually shared information. i was thinking of a UUID system for online
 };
-
 
 class BandGameplayStats : public PlayerGameplayStats {
 public:
@@ -243,6 +240,3 @@ public:
 
     void DrumNotePoint(bool perfect, int playerMult, bool cymbal);
 };
-
-
-
